Initial implementation of RLBP.
There are some performance improvements that should be implemented like
improving the logic inside process and remove the ternary operators.

The initial part of the algorithm calculates lookup tables to speed up the
conversion from a standard 256-bin histogram into a histogram mainly composed
by uniform patterns.

The main bulk of the algorithm computes the 256-bin histogram just like standard LBP.

The main difference between RLBP and LBP is found in the epilogue function.
For each bin in the standard 256-bin histogram we detect and replace arbitrary
bit patterns that are usually linked with noise with other more meaningful patterns.
These new patterns can be either uniform or non-uniform and the resulting
histogram must be updated accordingly.